requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8
redis==3.5
celery==4.4.7
flower==0.9.7
django-celery-beat==2.2.0
requests==2.25.1
djangorestframework==3.12.4